+++ sources/xcode/Extensions/JLContacts/JLContacts/JLContacts.js
@@ -1,0 +1,8 @@
+ (() => {
+     if (window && window.$contacts) {
+         return;
+     }
+     
+     window.$contacts = {};
+     window.$contacts.all = () => $agent.trigger("$contacts.all", {});
+ })();