+++ sources/xcode/Extensions/JLContacts/JLContactsTests/JLContactsTests.m
@@ -1,0 +1,36 @@
+//
+//  JLContactsTests.m
+//  JLContactsTests
+//
+//  Created by Camilo Castro on 05-02-23.
+//
+
+#import <XCTest/XCTest.h>
+
+@interface JLContactsTests : XCTestCase
+
+@end
+
+@implementation JLContactsTests
+
+- (void)setUp {
+    // Put setup code here. This method is called before the invocation of each test method in the class.
+}
+
+- (void)tearDown {
+    // Put teardown code here. This method is called after the invocation of each test method in the class.
+}
+
+- (void)testExample {
+    // This is an example of a functional test case.
+    // Use XCTAssert and related functions to verify your tests produce the correct results.
+}
+
+- (void)testPerformanceExample {
+    // This is an example of a performance test case.
+    [self measureBlock:^{
+        // Put the code you want to measure the time of here.
+    }];
+}
+
+@end